♻️ Split Feature/Group files, update tests and docs
- Split {Feature}Feature.cs into {Feature}Feature.cs + {Feature}Group.cs
  so each class has its own file.
- Point integration-test usings at the new per-slice namespaces.
- Add ADR 20260515 recording the folder-per-slice decision.
- Update AGENTS.md to describe the new structure.
