+++ src/WebApi/Features/Heroes/HeroesFeature.cs
@@ -9,12 +9,3 @@ public sealed class HeroesFeature : IFeature
         // TODO: Add feature-specific services here
     }
 }
-
-public class HeroesGroup : Group
-{
-    public HeroesGroup()
-    {
-        // NOTE: The prefix is used as the tag and group name
-        base.Configure("heroes", ep => ep.Description(x => x.ProducesProblemDetails(500)));
-    }
-}
