+++ src/WebApi/Features/Teams/TeamsFeature.cs
@@ -1,4 +1,4 @@
-﻿using SSW.VerticalSliceArchitecture.Common.Interfaces;
+using SSW.VerticalSliceArchitecture.Common.Interfaces;
 
 namespace SSW.VerticalSliceArchitecture.Features.Teams;
 
@@ -7,14 +7,5 @@ public sealed class TeamsFeature : IFeature
     public static void ConfigureServices(IServiceCollection services, IConfiguration config)
     {
         // TODO: Add feature-specific services here
-    }
-}
-
-public class TeamsGroup : Group
-{
-    public TeamsGroup()
-    {
-        // NOTE: The prefix is used as the tag and group name
-        base.Configure("teams", ep => ep.Description(x => x.ProducesProblemDetails(500)));
     }
 }
