+++ tests/WebApi.IntegrationTests/Endpoints/Heroes/Commands/CreateHeroCommandTests.cs
@@ -1,7 +1,7 @@
 using FastEndpoints;
 using Microsoft.EntityFrameworkCore;
 using SSW.VerticalSliceArchitecture.Common.Domain.Heroes;
-using SSW.VerticalSliceArchitecture.Features.Heroes.Endpoints;
+using SSW.VerticalSliceArchitecture.Features.Heroes.CreateHero;
 using SSW.VerticalSliceArchitecture.IntegrationTests.Common;
 using System.Net;
