+++ tests/WebApi.IntegrationTests/Endpoints/Teams/Events/UpdatePowerLevelEventTests.cs
@@ -3,7 +3,7 @@ using FastEndpoints;
 using Microsoft.EntityFrameworkCore;
 using SSW.VerticalSliceArchitecture.Common.Domain.Heroes;
 using SSW.VerticalSliceArchitecture.Common.Domain.Teams;
-using SSW.VerticalSliceArchitecture.Features.Heroes.Endpoints;
+using SSW.VerticalSliceArchitecture.Features.Heroes.UpdateHero;
 using SSW.VerticalSliceArchitecture.IntegrationTests.Common;
 using SSW.VerticalSliceArchitecture.IntegrationTests.Common.Factories;
 using SSW.VerticalSliceArchitecture.IntegrationTests.Common.Utilities;
